+++ fercord_storage/src/model/mod.rs
@@ -5,3 +5,6 @@ pub mod reminder;
 
 /// Store guild timezones as setting data
 pub mod guild_timezone;
+
+pub use reminder::*;
+pub use guild_timezone::*;